+++ src/main/java/org/hmanwon/domain/member/dao/MemberRepository.java
@@ -2,7 +2,9 @@ package org.hmanwon.domain.member.dao;
 
 import org.hmanwon.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
 }
