+++ src/main/java/org/hmanwon/domain/member/entity/Member.java
@@ -13,12 +13,13 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.validation.constraints.Min;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hmanwon.domain.community.board.entity.Board;
 import org.hmanwon.domain.community.comment.entity.Comment;
-import org.hmanwon.domain.zzan.entity.PurchaseHistory;
+import org.hmanwon.domain.zzan.purchase.entity.PurchaseHistory;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
@@ -26,6 +27,7 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class Member {
 
     @Id
@@ -44,6 +46,7 @@ public class Member {
     @Column(nullable = false)
     @org.hibernate.annotations.Comment("회원 포인트")
     @Min(0)
+    @Setter
     private Long point;
 
     @OneToMany(mappedBy = "member",
