+++ src/main/java/org/hmanwon/domain/shop/entity/SeoulGoodShop.java
@@ -20,7 +20,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.Comment;
-import org.hmanwon.domain.zzan.entity.ZzanItem;
+import org.hmanwon.domain.zzan.zzanItem.entity.ZzanItem;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
