+++ src/main/java/org/hmanwon/domain/shop/init/application/InitSeoulGoodShopService.java
@@ -13,8 +13,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import lombok.RequiredArgsConstructor;
@@ -31,9 +29,9 @@ import org.hmanwon.domain.shop.init.exception.InitShopException;
 import org.hmanwon.domain.shop.init.exception.InitShopExceptionCode;
 import org.hmanwon.domain.shop.init.util.AddressToCodeConverter;
 import org.hmanwon.domain.shop.init.util.KakaoApiUtil;
-import org.hmanwon.domain.zzan.dao.ZzanItemRepository;
-import org.hmanwon.domain.zzan.entity.ZzanItem;
-import org.hmanwon.domain.zzan.type.SaleStatus;
+import org.hmanwon.domain.zzan.zzanItem.dao.ZzanItemRepository;
+import org.hmanwon.domain.zzan.zzanItem.entity.ZzanItem;
+import org.hmanwon.domain.zzan.zzanItem.type.SaleStatus;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.stereotype.Service;
