--- src/main/java/org/hmanwon/domain/zzan/dao/PurchaseHistoryRepository.java
@@ -1,10 +1,0 @@
-package org.hmanwon.domain.zzan.dao;
-
-import org.hmanwon.domain.zzan.entity.PurchaseHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PurchaseHistoryRepository extends JpaRepository<PurchaseHistory, Long> {
-
-}
