+++ src/main/java/org/hmanwon/domain/zzan/purchase/type/PurchaseStatusType.java
@@ -1,4 +1,4 @@
-package org.hmanwon.domain.zzan.type;
+package org.hmanwon.domain.zzan.purchase.type;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum PurchaseStatusType {
 
-    IN_USE("사용했음"),
-    TO_BE_USE("구매는 했고, 사용은 예정")
+    IN_USE("사용 완료"),
+    TO_BE_USE("구매 완료, 사용 예정")
     ;
 
     private String type;
