+++ src/main/java/org/hmanwon/domain/zzan/zzanItem/exception/ZzanExceptionCode.java
@@ -1,4 +1,4 @@
-package org.hmanwon.domain.zzan.exception;
+package org.hmanwon.domain.zzan.zzanItem.exception;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
